Modify the ZkServerWrapper to reuse the same in-process zookeeper instance for each request
Each request gets an object were getAddress() resolves to host:port/chroot.  Chroots are generated
in bulk because establishing a connectiont to zookeeper is pretty slow.  Total performance gains are
moderate, ~20 second improvement in total test run time. (out of ~8 minutes)

RB=963036
G=voldemort-reviewers
A=goliu

+++ venice-test-common/src/main/java/com/linkedin/venice/integration/utils/ProcessWrapper.java
@@ -118,7 +118,7 @@ public abstract class ProcessWrapper implements Closeable {
       LOGGER.error("Failed to shutdown " + serviceName + " service running at " + getAddressForLogging(), e);
     }
     try {
-      if (dataDirectory != null) {
+      if (dataDirectory != null && dataDirectory.exists()) {
         FileUtils.deleteDirectory(dataDirectory);
       }
     } catch (IOException e) {
